Added a new links partial
Restyled the existing links for visual polish and accessibility and
pulled it all into a new separate link partial. Created a color mixin
that should be sustainable and defineable by block element. Currently,
this commit is defining link styling within

*`.content-main`
*`footer`
*`.govau—header`
*`hero`

Additionally, there is an independent class `.templink` created for
inactive link placeholders (which GOV.AU will have a LOT of in its beta)