Group of links, aka "Related" or "Popular" links to appear in heroes or at the ends of articles.
Messing with Bourbon Neat layout setup with ULs to get a multipurpose
group of links to appear in multiple columns. Primary focus is on the
group of links titled “Popular now” to appear in homepage heroes. It
should also be flexible to live in other areas, i.e. after an article