239 - Multipurpose group of links
Messing with Bourbon Neat layout setup with ULs to get 2-4 columns of
links to appear in. Meant to be multipurpose for use in various areas.

Primary focus is on the
group of links titled “Popular now” to appear in homepage heroes. It
should also be flexible to live in other areas, i.e. after an article