Group of links ("Related"/Popular")
## Description

Messing with Bourbon Neat layout setup with ULs to get 2-4 columns of
links to appear in. Meant to be multipurpose for use in various areas.

Primary focus is on the group of links titled “Popular now” to appear
in homepage heroes. It
should also be flexible to live in other areas, i.e. after an article

## Additional information

* I can't get the links into two columns in mobile. Got stuck with the
omega bug.
* Desktop and desktop seem alright staying in 4 columns, but hopefully
I'm not just settling because I got stuck...