Omega hack Pt. II
The hack worked to fix what I was trying to fix (columns showing up
correctly). There’s a bit of a rehash in the grid settings,
particularly some new variables to call out the mininum width of
breakpoints (as these numbers are used in various places)

Also cleaned up the example file to show things in the correct context.