Edits to make default link style apply to all content
Specific link styles to `.content-main` have been replaced so that
those link styles are applied to any `article` within `main`. Keeping
it more high level.

The `content-main` class, which the link style was originally applied
to, was specifically a content area for layouts with a sidebar…

That said…
NOTE: Should revisit the naming structure for `content-main`, and
clarify full-width layouts, layouts with sidebars, layouts that have
multiple columns within the `content-main` etc….