Change examples page title
More relevant file name.

After chatting to Donna I think this is ok from an IA perspective.

If this was of a more public page/site we would consider renaming the
filename to something more informative  without needing the rest of the
URL to make sense of it (eg `all-examples`), however as it’s not a
semantic issue I have opted to keep a short (but more relevant) file
name.

Likely we will return to the naming of this file with further IA work.